ENGR00241595-1 mx6q-sabresd:Enable SATA PHY PDDQ default
The SATA PHY PDDQ configuration is depended on the CONFIG_CMD_SATA
defined or not in board config file. If SATA feature is not config,
then the PDDQ will not be set, SATA PHY will not entry in Low Power
Mode, and it will consume some power even there is no sata devices
on board.
This patch:
1 Enable SATA PHY PDDQ default no matter the SATA is enabled or not in
  board config file, SATA module will disable PDDQ first when used phy,
  so default enable PDDQ will not affect SATA feature.
2 It needs a delay to wait for SATA PHY initialize after enable it,
  otherwise write the phy registers will fail.

Signed-off-by: Lin Fuzhen <fuzhen.lin@freescale.com>